warnings removed

+++ itomIsoGLFigurePlugin/plotIsoGLWidget.cpp
@@ -1406,14 +1406,14 @@ void plotGLWidget::refreshPlot(ito::ParamBase *param)
             int y1 = m_pContent->getSize(dims - 2) - 1;
             bool test;
 
-            if((m_axisX.autoScale) || (m_axisX.idx[1] > x1) || (m_axisX.dimIdx != dims - 1))
+            if((m_axisX.autoScale) || ((int)m_axisX.idx[1] > x1) || ((int)m_axisX.dimIdx != dims - 1))
             {
                 m_axisX.idx[0] = 0;
                 m_axisX.idx[1] = x1;
                 m_axisX.dimIdx = dims - 1;
             }
 
-            if((m_axisY.autoScale) || (m_axisY.idx[1] > y1) || (m_axisY.dimIdx != dims - 2))
+            if((m_axisY.autoScale) || ((int)m_axisY.idx[1] > y1) || ((int)m_axisY.dimIdx != dims - 2))
             {
                 m_axisY.idx[0] = 0;
                 m_axisY.idx[1] = y1;
@@ -2560,7 +2560,7 @@ void plotGLWidget::DrawObjectInfo(void)
     y0 -= 3.0 * m_fontsize/ height();
     if(m_objectInfo.matrix.length()) OGLTextOut((char*)m_objectInfo.matrix.data(), x0, y0);
 
-    int len = m_objectInfo.PeakText.length();
+    size_t len = m_objectInfo.PeakText.length();
 
     if(len < m_objectInfo.MeanText.length())
         len = m_objectInfo.MeanText.length();
